+++ tests/Pigeon.Messaging.Outbox.InMemory.Tests/InMemoryOutboxTestMessage.cs
@@ -1,0 +1,7 @@
+namespace Pigeon.Messaging.Outbox.InMemory.Tests
+{
+    internal sealed class InMemoryOutboxTestMessage
+    {
+        public string Text { get; set; }
+    }
+}